explicitly set type in interdf

+++ Vseq.py
@@ -797,7 +797,7 @@ def plotPpmMatrix(sub, plainseq, fppm, dm, frags, zoom, ions, err, specpar, exp_
     for column in fppm.T:
         if (fppm.T[column]<50).any():
             try:
-                interdf[column][int(column[1:])-1] = math.log(fragints[column])
+                interdf[column][int(column[1:])-1] = int(math.log(fragints[column]))
                 # intlist.append(math.log(fragints[column]))
             except KeyError:
                 interdf[column][int(column[1:])-1] = 0
